% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ssmRG.R
\name{calculate_ogive}
\alias{calculate_ogive}
\title{Calculate ogive}
\usage{
calculate_ogive(data, method = "fq", niter = 1000, seed = 70387)
}
\arguments{
\item{data}{object of class \code{classify} with the X, Y and mature classification (juvelines = 0, adults = 1).}

\item{method}{the method to be applied, \code{"fq"} frecuentist GLM, or \code{"bayes"} bayesian GLM
(MCMClogit function).}

\item{niter}{number of iterations (bootstrap resampling).}

\item{seed}{a single value, interpreted as an integer.}
}
\value{
object of class \code{ogive} with the parameters, the L50 and data.frame with the X, Y and mature classification
variables. Also the fitted values for the logistic regression and confidence intervals.
}
\description{
Estimate the size at maturity using a logistic regression
relating X variable and maturity stage (classifying individuals as juveniles 
or adults depending on their morphometry).
}
\examples{
my_file = system.file("extdata", "crabdat.txt", package = "ssmRG")
data    = read_data(my_file)
classify_data = classify_mature(data, sex = "all")
my_ogive = calculate_ogive(classify_data, method = "fq")
my_ogive$A_boot
my_ogive$B_boot
my_ogive$L50_boot
my_ogive$out
}

