% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crabdata}
\alias{crabdata}
\title{Crab data}
\format{A data frame with 3 variables: \code{carapace_width} in mm, 
\code{chela_heigth} in mm and \code{sex_category} (m: male, f: female).}
\usage{
crabdata
}
\description{
A data set containing measures and other atribbutes of almost 223 crabs.
}
\keyword{datasets}

