% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphMat-main.R
\name{morph_mature}
\alias{morph_mature}
\title{Calculate morphometric mature}
\usage{
morph_mature(data, method = "fq", niter = 999, seed = 70387)
}
\arguments{
\item{data}{an object of class 'classify' with the allometric variables (X", "Y") and classification of maturity (juvelines = 0, adults = 1).}

\item{method}{a character string indicating the method to be applied, \code{"fq"} frecuentist GLM, or \code{"bayes"} bayesian GLM (MCMClogit function).}

\item{niter}{number of iterations (bootstrap resampling).}

\item{seed}{a single value, interpreted as an integer.}
}
\value{
object of class 'morphMat'.

\code{model} the summary model.

\code{A_boot} the 'n iter' values of parameter A.

\code{B_boot} the 'n iter' values of parameter B.

\code{L50} the 'n iter' values of parameter L50 (morphometric size at sexual maturity).

\code{out} a dataframe with the allometric variables "X" and "Y", classification of maturity, the fitted values for  
logistic regression and confidence intervals (95\%). Also the summary model is provided.
}
\description{
Estimate the morphometric size at sexual maturity (\eqn{L50}).
}
\details{
Estimate the morphometric size at sexual maturity using a logit regression with X variable 
and maturity classification (two categories: juveniles and adults). 

The function requires an object of class "classify" with the X, Y (allometric variables) and classification of maturity (juvelines = 0, adults = 1).

The argument `method` requires a character string indicanting which regression will be used for the test.
If `method = "fq"` the regression is based on GLM (frequentist) and if `method = "bayes"` a sample from the posterior distribution 
of a logistic regression model using a random walk Metropolis algorithm is generated (see MCMClogit function).

The argument `niter` requires a number. For the GLM regression (`method = "fq"`), a non-parametric bootstrap method consists
in generate B bootstrap samples, by resampling with replacement the original data. Then all statistics for each parameter 
can be calculated from each bootstrap sample (median and confidence intervals). 
For the `method = "bayes"`, the argument `niter` is related to the number of Metropolis iterations for the sampler.
}
\examples{
\dontrun{
data(crabdata)

classify_data = classify_mature(crabdata, varnames = c("carapace_width", "chela_heigth"), 
varsex = "sex_category", selectSex = NULL, method = "ld")

my_mature = morph_mature(classify_data, method = "fq")

## \\eqn{niter} parameters
my_mature$A_boot
my_mature$B_boot
my_mature$L50_boot
my_mature$out
}
}

