% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphMat-main.R
\name{plot.classify}
\alias{plot.classify}
\title{Plot method for classify class}
\usage{
\method{plot}{classify}(
  x,
  xlab = "X",
  ylab = "Y",
  col = c(1, 2),
  pch = c(4, 5),
  cex = c(1, 1),
  lty_lines = c(1, 1),
  lwd_lines = c(1, 1),
  legendPlot = TRUE,
  cex_label = 0.8,
  gg_style = FALSE,
  point_alpha = 0.8,
  base_size = 13,
  ...
)
}
\arguments{
\item{x}{an object of class 'classify' with the allometric variables ("X", "Y") and classification of maturity (juveniles = 0, adults = 1).}

\item{xlab}{title for the x axis.}

\item{ylab}{title for the y axis.}

\item{col}{colors for juveniles and adults group.}

\item{pch}{plotting characters for juveniles and adults.}

\item{cex}{character expansion for points.}

\item{lty_lines}{line types for the regression lines.}

\item{lwd_lines}{line widths for the regression lines.}

\item{legendPlot}{legend in the plot (FALSE or TRUE).}

\item{cex_label}{size of the legend text in base graphics.}

\item{gg_style}{logical. If TRUE, return a ggplot2 object.}

\item{point_alpha}{transparency level of points, used only when gg_style = TRUE.}

\item{base_size}{base font size, used only when gg_style = TRUE.}

\item{\dots}{additional arguments passed to the base plot method.}
}
\description{
Plot method for classify class
}
\examples{
data(crabdata)

classify_data = classify_mature(crabdata, varNames = c("carapace_width", "chela_height"),
varSex = "sex_category", selectSex = NULL, method = "ld")

## Showing different plots
plot(classify_data, xlab = "X")

plot(classify_data, xlab = "X", ylab = "Y", col = c(1, 2), pch = c(4, 5), cex = c(1, 3))

plot(classify_data, xlab = "Carapace width (mm.)", ylab = "Y", col = c(1, 2),
pch = c(4, 5), cex = c(1, 3), lwd_lines = c(1, 3))

plot(classify_data, xlab = "Carapace width (mm.)", ylab = "Y", col = c(1, 2),
pch = c(4, 5), cex = c(1, 3), lwd_lines = c(1, 3), main = "Classification")
}
