% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphMat-main.R
\name{print.classify}
\alias{print.classify}
\title{Print method for classify class}
\usage{
\method{print}{classify}(x, ...)
}
\arguments{
\item{x}{an object of class 'classify' with the allometric variables ("X", "Y") and classification of maturity (juveniles = 0, adults = 1).}

\item{\dots}{Additional arguments to the print method.}
}
\value{
The number of juveniles and adults. Also shows the regression analysis for juveniles and adults
and an ANCOVA analysis to compare slopes.
}
\description{
Print method for classify class
}
\examples{
data(crabdata)

classify_data = classify_mature(crabdata, varNames = c("carapace_width", "chela_height"),
varSex = "sex_category", selectSex = NULL, method = "ld")

print(classify_data)
}
