% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gonadMat-main.R
\name{print.gonadMat}
\alias{print.gonadMat}
\title{Print method for gonadMat class (size at gonad maturity)}
\usage{
\method{print}{gonadMat}(x, ...)
}
\arguments{
\item{x}{object of class 'gonadMat' with the parameters of the logistic regression and a data.frame with the X and stage of sexual maturity.
variables. Also the fitted values for the logistic regression and confidence intervals (95\%).}

\item{\dots}{Additional arguments to the print method.}
}
\value{
The median of the size at gonad maturity estimation, parameters and the Nagelkerke's R squared.
}
\description{
Print method for gonadMat class (size at gonad maturity)
}
\examples{
data(matFish)

gonad_mat = gonad_mature(matFish, varNames = c("total_length", "stage_mat"), inmName = "I", 
matName = c("II", "III", "IV"), method = "fq", niter = 50)

print(gonad_mat)
}
