% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ssmRG.R
\name{print.mature}
\alias{print.mature}
\title{Print mature database}
\usage{
\method{print}{mature}(x, ...)
}
\arguments{
\item{x}{object of class \code{mature}.}

\item{\dots}{Additional arguments to be passed}
}
\description{
Print mature database
}
\examples{
my_file = system.file("extdata", "crabdat.txt", package = "ssmRG")
data    = read_data(my_file)
print(data)
}

