% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphMat-main.R
\name{print.morphMat}
\alias{print.morphMat}
\title{Print morphometric size at sexual maturity}
\usage{
\method{print}{morphMat}(x, ...)
}
\arguments{
\item{x}{object of class 'morphMat' with the mature parameters and a data.frame with the allometric variables ("X", "Y") 
and classification of maturity. Also the fitted values for the logistic regression and confidence intervals (95\%).}

\item{\dots}{Additional arguments to the print method.}
}
\description{
Print morphometric size at sexual maturity
}
\examples{
\dontrun{
data(crabdata)

classify_data = classify_mature(crabdata, varnames = c("carapace_width", "chela_heigth"), 
varsex = "sex_category", selectSex = NULL, method = "ld")

my_mature = morph_mature(classify_data, method = "fq")

print(my_mature)
}
}

