% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ssmRG.R
\name{print.ogive}
\alias{print.ogive}
\title{Print maturity ogive}
\usage{
\method{print}{ogive}(x, probs = c(0.025, 0.5, 0.975), ...)
}
\arguments{
\item{x}{object of class 'ogive' with the ogive parameters and a data.frame with the X, Y and mature classification
variables. Also the fitted values for the logistic regression and confidence intervals.}

\item{probs}{numeric vector of probabilities with values in \code{[0,1]}.}

\item{\dots}{Additional arguments to the print method.}
}
\description{
Print maturity ogive
}
\examples{
data(crabdata)
classify_data = classify_mature(crabdata, varnames = c("carapace_width", "chela_heigth"), 
varsex = "sex_category", selectSex = NULL, method = "ld")
classify_data
my_ogive = calculate_ogive(classify_data, method = "fq")
print(my_ogive)
}

