% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ssmRG.R
\name{read_data}
\alias{read_data}
\title{Read data}
\usage{
read_data(file)
}
\arguments{
\item{file}{the filename (extension: txt, csv, csv2).}
}
\value{
an object of class \code{mature} with x (independent), y (dependent) variables and sex category (\code{m} = male,
\code{f} = female, \code{ind} = indeterminate sex).
}
\description{
Read a database with different extesions. The database has three variables, "x", "y" and "sex".
The variables "x" and "y" will be used in the regression. The variable "sex" 
is the sex category (\code{m} = male, \code{f} = female, \code{ind} = indeterminate sex.)
}
\examples{
my_file = system.file("extdata", "crabdat.txt", package = "ssmRG")
data = read_data(my_file)
}

