\docType{package}
\name{ssmRG-package}
\alias{ssmRG-package}
\alias{ssmRG}
\title{Size at Sexual Maturity from morphometric data}
\description{
Estimate size at sexual maturity from morphometic data based on relative growth (alometric growth). The classification use PCA, hierarchical cluster and linear or quadratic regression analysis.
}
\details{
\tabular{ll}{
Package: \tab ssmRG\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-06-12\cr
License: \tab GPL-2\cr
}

\author{
Edgar Josymar Torrejon-Magallanes <ejosymart@gmail.com>
}

\references{
Corgos, A., and Freire, J. (2006). Morphometric and gonad maturity in the spider crab Maja brachydactyla: a comparison of methods for estimating size at maturity in species with determinate growth. ICES Journal of Marine Science: Journal du Conseil, 63(5), 851-859.

Sampedro, M. P., Gonzalez-Gurriaran, E., Freire, J., and Muino, R. (1999). Morphometry and sexual maturity in the spider crab Maja squinado (Decapoda: Majidae) in Galicia, Spain. Journal of Crustacean Biology, 578-592.

Somerton, D. A. (1980). A computer technique for estimating the size of sexual maturity in crabs. Canadian Journal of Fisheries and Aquatic Sciences, 37(10), 1488-1494.

https://github.com/petrkeil/bPCA
}
\keyword{package}