% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ssmRG.R
\name{calculate_ogive}
\alias{calculate_ogive}
\title{Calculate ogive}
\usage{
calculate_ogive(data, method = "fq")
}
\arguments{
\item{data}{The database with the X, Y and mature stages (juvelines = 0, adults = 1)}

\item{method}{The method to be applied, "fq" frecuentist GLM, or "bayes" bayesian GLM
(MCMClogit function).}
}
\value{
list Database with the parameters and a data.frame with the X, Y and mature stages
variables. Also the fitted values for the logistic regression and confidence intervals.
}
\description{
Estimate the size at maturity using a logistic regression
relating X variable and maturity stage (classifying individuals as juveniles 
or adults depending on their morphometry).
}
\examples{
my_file = system.file("extdata", "crabdat.txt", package = "ssmRG")
data    = read_data(my_file)
classify_data = classify_mature(data)
my_ogive = calculate_ogive(classify_data, method = "fq")
}

