% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ssmRG.R
\name{calculate_ogive}
\alias{calculate_ogive}
\title{Calculate ogive}
\usage{
calculate_ogive(data, method = "fq", niter = 1000, seed = 70387)
}
\arguments{
\item{data}{an object of class 'classify' with the X, Y and mature classification (juvelines = 0, adults = 1).}

\item{method}{the method to be applied, \code{"fq"} frecuentist GLM, or \code{"bayes"} bayesian GLM (MCMClogit function).}

\item{niter}{number of iterations (bootstrap resampling).}

\item{seed}{a single value, interpreted as an integer.}
}
\value{
object of class 'ogive'.

\code{model} the summary model.

\code{A_boot} the 'n iter'values of parameter A.

\code{B_boot} the 'n iter'values of parameter B.

\code{L50} the 'n iter'values of parameter L50 (size at sexual maturity).

\code{out} a dataframe with the X, Y and mature classification variables. 
Also the fitted values for the logistic regression and confidence intervals (95\%).
}
\description{
Estimate the size at sexual maturity (L50).
}
\details{
Estimate the size at sexual maturity using a logit regression with X variable and maturity stage (two categories: juveniles 
and adults).
}
\examples{
data(crabdata)
classify_data = classify_mature(crabdata, varnames = c("carapace_width", "chela_heigth"), 
varsex = "sex_category", useSex = NULL, method = "ld")
classify_data
my_ogive = calculate_ogive(classify_data, method = "fq")
my_ogive$A_boot
my_ogive$B_boot
my_ogive$L50_boot
my_ogive$out
}

