% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ssmRG.R
\name{calculate_ogive}
\alias{calculate_ogive}
\title{Calculate ogive}
\usage{
calculate_ogive(data, methodReg = "fq")
}
\arguments{
\item{data}{The database with the X, Y and mature stages (juvelines = 0, adults = 1)}

\item{methodReg}{The method to be applied, "fq"frecuentist GLM, or "bayes" bayesian GLM
(MCMClogit function).}
}
\value{
List database with the parameters and a data.frame with the X, Y and mature stages
variables. Also the fitted values for the logistic regression and confidence intervals.
}
\description{
Estimate the size at 50\% maturity using a logistic regression
relating X variable and maturity stage (classifying individuals into juveniles 
or adults depending on their morphometry).
}
\examples{
my_ogive = calculate_ogive(data, methodReg = "fq")
plot(my_ogive, xlab = "X", ylab = "Proportion mature", col = "blue", col50 = "red")
}

