% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ssmRG.R
\name{classify_mature}
\alias{classify_mature}
\title{Classify mature}
\usage{
classify_mature(data)
}
\arguments{
\item{data}{The database with two variables (x: independent variable, y: dependent variable)}
}
\value{
Database with x (independent), y (dependent) and mature classification
(juveniles = 0, adult = 1) variables.
}
\description{
Classify in two groups (juvelines = 0 and adult = 1). The analisys is based on 
Principal Components Analisys with the variables 
(x: independent variable, y: dependent variable) in log base, allowing to distinguish 
two groups would represent juveniles and adult.
The individuals are assigned to each group using a hierarchical classification 
procedure (hierarchical cluster). This method is based on establishing a predetermined 
number of groups (in this case, two) and assigning individuals to one of the groups according 
to their loads on the two axes of the PCA.
Using the results of the classification (PCA + cluster), a discriminant analysis (linear or quadratic) 
is conducted to obtain a discriminating function that permitted any individuals 
to be classified as a juvenile or an adult on the basis of the X and Y variables.
}
\examples{
my_file = system.file("extdata", "crabdat.txt", package = "ssmRG")
data    = read_data(my_file)
classify_data = classify_mature(data)
classify_data
}

