% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crabdata}
\alias{crabdata}
\title{Crab data}
\format{A data frame with different variables:

\code{year}

\code{month}

\code{carapace_width} (mm.)

\code{carapace_length} (mm.)

\code{chela_heigth} (mm.)

\code{chela_width} (mm.)
 
\code{sex_category} (m: male, f: female).}
\source{
Somerton, D. A., & Donaldson, W. (1996). Contribution of the biology of the grooved and triangle Tanner crabs,
Chionoecetes tanneri and C. angulatus, in the eastern Bering Sea. Fishery Bulletin, 94(2), 348-357.
}
\usage{
crabdata
}
\description{
A data set containing allometric measures and other atribbutes of 223 crabs of the specie \eqn{Chionoectes tanneri}.
The variables carapace width and chela heigth were published in Somerton & Donaldson (1996).
}
\references{
Watters, G., & Hobday, A. J. (1998). A new method for estimating the morphometric size at maturity of crabs. 
Canadian Journal of Fisheries and Aquatic Sciences, 55(3), 704-714.
}
\keyword{datasets}

