% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fish maturity}
\alias{Fish maturity}
\alias{matFish}
\title{Fish gonadal maturity data}
\format{A data frame with two variables:

\code{total_length} the total lenght in cm.

\code{stage_mat} the stages of sexual maturity (I: inmature, II-III-IV: mature)}
\usage{
data(matFish)
}
\description{
A data set containing the total length and the stages of sexua maturity of 1733 fishes.
}

