% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmRG-main.R
\name{plot.classify}
\alias{plot.classify}
\title{Plot classify data}
\usage{
\method{plot}{classify}(x, xlab = "X", ylab = "Y", col = c(1, 2),
  pch = c(4, 5), cex = c(1, 1), lty_lines = c(1, 1), lwd_lines = c(1,
  1), ...)
}
\arguments{
\item{x}{an object of class 'classify', with x (independent), y (dependent) and mature classification (juveniles = 0, adult = 1) variables.}

\item{xlab}{a title for the x axis.}

\item{ylab}{a title for the y axis.}

\item{col}{the colors for juveniles and adults group.}

\item{pch}{the character indicating the type of plotting.}

\item{cex}{character expansion in the regression.}

\item{lty_lines}{the line type in the regression.}

\item{lwd_lines}{the line width in the regression.}

\item{\dots}{Additional arguments to the plot method.}
}
\description{
Plot classify data
}
\examples{
data(crabdata)

classify_data = classify_mature(crabdata, varnames = c("carapace_width", "chela_heigth"), 
varsex = "sex_category", selectSex = NULL, method = "ld")

plot(classify_data, xlab = "X")

plot(classify_data, xlab = "X", ylab = "Y", col = c(1, 2), pch = c(4, 5), cex = c(1, 3))

plot(classify_data, xlab = "Carapace width (mm.)", ylab = "Y", col = c(1, 2), 
pch = c(4, 5), cex = c(1, 3), lwd_lines = c(1, 3))

plot(classify_data, xlab = "Carapace width (mm.)", ylab = "Y", col = c(1, 2), 
pch = c(4, 5), cex = c(1, 3), lwd_lines = c(1, 3), main = "Classification")

}

