% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ssmRG.R
\name{plot.classify}
\alias{plot.classify}
\title{Plot classify data}
\usage{
\method{plot}{classify}(x, xlab = "X", ylab = "Y", col = c(1, 2),
  pch = c(4, 5), ...)
}
\arguments{
\item{x}{Database (a list) with x (independent), y (dependent) and mature classification (juveniles = 0, adult = 1) variables.}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{col}{the colors for juveniles and adults group}

\item{pch}{the character indicating the type of plotting}

\item{\dots}{Additional arguments to be passed to methods}
}
\description{
Plot classify data
}
\examples{
my_file = system.file("extdata", "crabdat.txt", package = "ssmRG")
data    = read_data(my_file)
classify_data = classify_mature(data)
plot(classify_data, xlab = "X", ylab = "Y", col = c(1, 2), pch = c(4, 5))
}

