% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ssmRG.R
\name{plot.ogive}
\alias{plot.ogive}
\title{Plot maturity ogive}
\usage{
\method{plot}{ogive}(x, xlab1 = "X", ylab1 = "Proportion mature",
  col = c("blue", "red"), ...)
}
\arguments{
\item{x}{List database with the parameters and a data.frame with the X, Y and mature stages
variables. Also the fitted values for the logistic regression and confidence intervals.}

\item{xlab1}{a title for the x axis}

\item{ylab1}{a title for the y axis}

\item{col}{the color for the logistic curve for the 50\% size at sexual maturity}

\item{\dots}{Additional arguments to be passed to methods}
}
\description{
Plot maturity ogive
}
\examples{
my_file = system.file("extdata", "crabdat.txt", package = "ssmRG")
data    = read_data(my_file)
classify_data = classify_mature(data)
my_ogive = calculate_ogive(classify_data, method = "fq")
plot(my_ogive, xlab1 = "X", ylab1 = "Proportion mature", col = c("blue", "red"))
}

