% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gonadMat-main.R
\name{print.gonadMat}
\alias{print.gonadMat}
\title{Print gonadal size at sexual maturity}
\usage{
\method{print}{gonadMat}(x, ...)
}
\arguments{
\item{x}{object of class 'gonadMat' with the mature parameters and a data.frame with the X and stage of sexual maturity.
variables. Also the fitted values for the logistic regression and confidence intervals (95\%).}

\item{\dots}{Additional arguments to the print method.}
}
\description{
Print gonadal size at sexual maturity
}
\examples{
\dontrun{
data(matFish)

gonad_mat = gonad_mature(matFish, varNames = c("total_length", "stage_mat"), inmName = "I", 
matName = c("II", "III", "IV"), method = "fq", niter = 999)

print(gonad_mat)
}
}

