% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmRG-main.R
\name{print.mature}
\alias{print.mature}
\title{Print morphometric size at sexual maturity}
\usage{
\method{print}{mature}(x, probs = c(0.025, 0.5, 0.975), ...)
}
\arguments{
\item{x}{object of class 'mature' with the mature parameters and a data.frame with the X, Y and mature classification
variables. Also the fitted values for the logistic regression and confidence intervals.}

\item{probs}{numeric vector of probabilities with values in \code{[0,1]}.}

\item{\dots}{Additional arguments to the print method.}
}
\description{
Print morphometric size at sexual maturity
}
\examples{
\dontrun{
data(crabdata)

classify_data = classify_mature(crabdata, varnames = c("carapace_width", "chela_heigth"), 
varsex = "sex_category", selectSex = NULL, method = "ld")

my_mature = calculate_mature(classify_data, method = "fq")

print(my_mature)
}
}

