% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ssmRG.R
\name{print.ogive}
\alias{print.ogive}
\title{Print maturity ogive}
\usage{
\method{print}{ogive}(x, probs = c(0.05, 0.5, 0.95), ...)
}
\arguments{
\item{x}{object of class \code{ogive} with the ogive parameters and a data.frame with the X, Y and mature classification
variables. Also the fitted values for the logistic regression and confidence intervals.}

\item{probs}{numeric vector of probabilities with values in \code{[0,1]}.}

\item{\dots}{Additional arguments to be passed}
}
\description{
Print maturity ogive
}
\examples{
my_file = system.file("extdata", "crabdat.txt", package = "ssmRG")
data    = read_data(my_file)
classify_data = classify_mature(data, sex = "all")
my_ogive = calculate_ogive(classify_data, method = "fq")
print(my_ogive)
}

