% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ssmRG.R
\name{read_data}
\alias{read_data}
\title{Read data}
\usage{
read_data(file)
}
\arguments{
\item{file}{The filename (extension: txt, csv, csv2)}
}
\value{
Database with x (independent) and y (dependent) variables.
}
\description{
Read a database with different extesions. The data base have only two variables, first variable is the independet variable and the second is the dependent variable.
}
\examples{
my_file = system.file("extdata", "crabdat.txt", package = "ssmRG")
data = read_data(my_file)
}

