% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ssm.R
\name{read_data}
\alias{read_data}
\title{Read data}
\usage{
read_data(file, ext = "txt")
}
\arguments{
\item{file}{The filename}

\item{ext}{The file extension ("txt", "csv")}
}
\value{
Database with x (independent) and y (dependent) variables.
}
\description{
Read a database with different extesions. The data base have only two variables, first variable is the independet variable and the second is the dependent variable.
}
\examples{
data = read_data("my_data", ext = "csv")
data
}

