% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ssmRG.R
\docType{package}
\name{ssmRG-package}
\alias{ssmRG-package}
\alias{ssmRG.}
\title{Size at Sexual Maturity based on Relative Growth.}
\description{
Package to estimate size at sexual maturity from morphometic data based on relative growth.
Some basic plotting (classification and maturity ogive) are also provided.
}
\details{
Package: ssmRG

Type: Package
}
\author{
Edgar Josymar Torrejon-Magallanes Maintainer: Edgar Josymar Torrejon-Magallanes
<ejosymart@gmail.com>
}
\references{
ssmRG: Size at Sexual Maturity based on Relative Growth (RJournal)
}
\keyword{-}
\keyword{alometric,}
\keyword{maturity,}
\keyword{relative-growth.}
\keyword{sexual}
\keyword{size,}

