% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmRG-main.R
\docType{package}
\name{ssmRG-package}
\alias{ssmRG-package}
\alias{ssmRG.}
\title{Morphometric Size at Sexual Maturity based on Relative Growth.}
\description{
Package to estimate morphometric size at sexual maturity based on relative growth.
}
\details{
Package: ssmRG

Type: Package

The estimation involves two process:

1) A Principal Components Analisys is conducted with two allometric variables (x: independent variable, y: dependent variable) in log base, allowing to distinguish 
two groups that would represent juveniles and adult. The individuals are assigned to each group using a hierarchical classification procedure (hierarchical cluster). 
This method is based on establishing a predetermined number of groups (in this case, two) and assigning individuals to one of the groups according to 
their loads on the two axes of the PCA (Corgos & Freire, 2006). Using the results of the classification (PCA + cluster), a discriminant analysis (linear or quadratic) 
is carried out to obtain a discriminating function that permitted any individuals to be classified as a juvenile or an adult on the basis of the X and Y 
allometric variables.

2) After classification, the logistic approach is used. The morphometric size at 50\% maturity (\eqn{L_50}) is estimated as the length at
which a randomly chosen specimen has a 50\% chance of being mature (Somerton  1980, Roa  et al. 1999, Corgos & Freire 2006). 
In the regression analysis, \eqn{X} (e.g: carapace width) is considered the explanatory variable and the classification stage \eqn{CS} 
(juvelines: 0, adults: 1) is considered the response variable (binomial). 
The variables are fitted to a logistic function with the form: 

\deqn{P_CS = 1 / [1+e^-(beta_0 + beta_1*X)]}

where:

\eqn{P_CL} is the probability of an individual of being mature at a determinate \eqn{X} length.
 
\eqn{beta_0} (intercept) and \eqn{beta_1} (slope) are parameters estimated. 

The (\eqn{L_50}) is calculated as: 

\deqn{L_50 = -beta_0 / beta_1}

Some basic plotting (classification, \eqn{beta_0}, \eqn{beta_1} and \eqn{L_50} histogram, and maturity ogive) 
are also provided.
}
\author{
Edgar Josymar Torrejon-Magallanes <ejosymart@gmail.com>
}
\references{
Agostinho, C. S. (2000). Use of otoliths to estimate size at sexual maturity in fish. Brazilian Archives of Biology and Technology, 43(4).

Corgos, A. & Freire, J. (2006). Morphometric and gonad maturity in the spider crab Maja brachydactyla: a comparison of methods for estimating size at maturity in species with determinate growth. ICES Journal of Marine Science: Journal du Conseil, 63(5), 851-859.

Roa, R., Ernst, B. & Tapia, F. (1999). Estimation of size at sexual maturity: an evaluation of analytical and resampling procedures. Fishery Bulletin, 97(3), 570-580.

Somerton, D. A. (1980). A computer technique for estimating the size of sexual maturity in crabs. Canadian Journal of Fisheries and Aquatic Sciences, 37(10), 1488-1494.
}
\keyword{alometric,}
\keyword{morphometric,}
\keyword{relative-growth.}
\keyword{sexual-maturity,}
\keyword{size,}

